part of merge

+++ OneStopAway/app/src/main/java/com/example/onestopaway/DataRepository.kt
@@ -76,7 +76,7 @@ class DataRepository(private val database :DatabaseManager) {
             spt = ln.split(",")
             numTripsFetched += 1
 
-            database.insertTrip(spt[10], spt[6], 0)
+            database.insertTrip(spt[10].toInt(), spt[6], 0)
         }
     }
 
@@ -95,7 +95,7 @@ class DataRepository(private val database :DatabaseManager) {
             split = line.split(",")
             numRoutesFetched += 1
 
-            database.insertRoute(split[0], split[1], split[2], split[3].toInt())
+            database.insertRoute(split[0].toInt(), split[1], split[2], split[3].toInt())
         }
         database.close()
     }
